docs: align docs with `FeedbackDataset` refactor using `tab-set`  (#3531)
This PR adds the `tab-set` for the outdated code-blocks with the
upcoming release of Argilla v1.14.0 so as to showcase both the previous
and the new code-blocks for users that are still using Argilla v1.14.0
or lower.

Additionally, the style has been fixed in some documents while reviewing
the outdated code-blocks, some file paths references have been fixed to
point to the HTML files via relative paths, and some minor details.

Finally, a `warning` has also been included to let the users know that
Argilla v1.14.0 won't work for the moment with the
`ArgillaCallbackHandler` in `LangChain`.

**Type of change**

- [X] Documentation update

---------

Co-authored-by: Gabriel Martín Blázquez <gmartinbdev@gmail.com>
Co-authored-by: Natalia Elvira <natalia@argilla.io>